Refactor directory and add improvements
