+++ game/src/snooker2D/colors/Brown.java
@@ -1,0 +1,9 @@
+package snooker2D.colors;
+
+import java.awt.*;
+
+public class Brown extends Color {
+    public Brown(){
+        super(139,69,19);
+    }
+}
