+++ game/src/snooker2D/colors/GameColors.java
@@ -1,0 +1,18 @@
+package snooker2D.colors;
+
+import java.awt.*;
+
+public class GameColors {
+
+    public static Color brown(){
+        return new Color(139,69,19);
+    }
+
+    public static Color pink(){
+        return new Color(255, 105,180);
+    }
+
+    public static Color cloth(){
+        return new Color(39, 174, 96);
+    }
+}
