+++ game/src/snooker2D/colors/Pink.java
@@ -1,0 +1,9 @@
+package snooker2D.colors;
+
+import java.awt.*;
+
+public class Pink extends Color {
+    public Pink(){
+        super(255, 105,180);
+    }
+}
